test-qcdm-com: fix warnings with -Wdouble-promotion
  test-qcdm-com.c: In function ‘test_com_pilot_sets’:
  test-qcdm-com.c:1085:59: warning: implicit conversion from ‘float’ to ‘double’ when passing argument to function [-Wdouble-promotion]
   1085 |         g_message ("      EC/IO     %d  (%.1f dB)", ecio, db);
        |                                                           ^~
  /usr/include/glib-2.0/glib/gmessages.h:333:32: note: in definition of macro ‘g_message’
    333 |                                __VA_ARGS__)
        |                                ^~~~~~~~~~~

+++ libqcdm/tests/test-qcdm-com.c
@@ -1082,7 +1082,7 @@ test_com_pilot_sets (void *f, void *data)
                                               &ecio,
                                               &db);
         g_message ("   %d: PN offset %d", i, pn_offset);
-        g_message ("      EC/IO     %d  (%.1f dB)", ecio, db);
+        g_message ("      EC/IO     %d  (%.1lf dB)", ecio, (double)db);
     }
 
     num = 0;
